requirements.txt
torch
torchvision
transformers
pandas
numpy
pillow
scikit-learn
tqdm
requests
